requirements.txt
# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
shap>=0.43.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
rioxarray>=0.15.0
xarray>=2023.1.0

# Google Earth Engine
earthengine-api>=0.1.380

# ECMWF data access
ecmwf-opendata>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data storage
pyarrow>=14.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Testing
pytest>=7.4.0
